[refactor] #211: 푸시 토큰 갱신 로직을 ArchiveMainViewModel에서 MainViewModel로 이동
